# Simplify and optimize isprime_I