--- a/src/core/interp.c
+++ b/src/core/interp.c
@@ -3538,9 +3538,8 @@ void MVM_interp_run(MVMThreadContext *tc, void (*initial_invoke)(MVMThreadContex
             }
             OP(isprime_I): {
                 MVMObject *a = GET_REG(cur_op, 2).o;
-                MVMint64 b = GET_REG(cur_op, 4).i64;
-                GET_REG(cur_op, 0).i64 = MVM_bigint_is_prime(tc, a, b);
-                cur_op += 6;
+                GET_REG(cur_op, 0).i64 = MVM_bigint_is_prime(tc, a);
+                cur_op += 4;
                 goto NEXT;
             }
             OP(rand_I): {