--- a/src/core/ops.c
+++ b/src/core/ops.c
@@ -6290,7 +6290,7 @@ static const MVMOpInfo MVM_op_infos[] = {
     {
         MVM_OP_isprime_I,
         "isprime_I",
-        3,
+        2,
         1,
         0,
         0,
@@ -6299,7 +6299,7 @@ static const MVMOpInfo MVM_op_infos[] = {
         0,
         0,
         0,
-        { MVM_operand_write_reg | MVM_operand_int64, MVM_operand_read_reg | MVM_operand_obj, MVM_operand_read_reg | MVM_operand_int64 }
+        { MVM_operand_write_reg | MVM_operand_int64, MVM_operand_read_reg | MVM_operand_obj }
     },
     {
         MVM_OP_rand_I,