--- a/src/math/bigintops.h
+++ b/src/math/bigintops.h
@@ -35,7 +35,7 @@ MVMnum64 MVM_bigint_to_num(MVMThreadContext *tc, MVMObject *a);
 MVMObject *MVM_bigint_from_num(MVMThreadContext *tc, MVMObject *result_type, MVMnum64 n);
 MVMnum64 MVM_bigint_div_num(MVMThreadContext *tc, MVMObject *a, MVMObject *b);
 MVMObject * MVM_bigint_rand(MVMThreadContext *tc, MVMObject *a, MVMObject *b);
-MVMint64 MVM_bigint_is_prime(MVMThreadContext *tc, MVMObject *a, MVMint64 b);
+MVMint64 MVM_bigint_is_prime(MVMThreadContext *tc, MVMObject *a);
 MVMObject * MVM_bigint_radix(MVMThreadContext *tc, MVMint64 radix, MVMString *str, MVMint64 offset, MVMint64 flag, MVMObject *type);
 MVMint64 MVM_bigint_is_big(MVMThreadContext *tc, MVMObject *a);
 MVMint64 MVM_bigint_bool(MVMThreadContext *tc, MVMObject *a);