… lib/MAST/Ops.nqp