Remove the `rounds` option. Rakudo only ever used a single (too large)
value and the op does not appear to be used in the ecosystem. Instead,
hardcode the value passed to the libtommath function to 40, chosen based
on https://security.stackexchange.com/questions/4544/how-many-iterations-of-rabin-miller-should-be-used-to-generate-cryptographic-saf
Also, libtommath now returns false for 1, so remove that check in our
code.

Additionally, add a branch for !MVM_BIGINT_IS_BIG, and handle that case
with the FJ32_256 algorithm from Forišek and Jančina, "Fast Primality
Testing for Integers That Fit into a Machine Word", 2015.

After making the single necessary modification to Rakudo,
`say (^100_000).grep(*.is-prime).tail; say now - INIT now` goes from
~12.8s to ~0.036s. The speedup isn't as dramatic for big integers,
`say (10_000_000_000..10_000_100_000).grep(*.is-prime).tail; say now - INIT now`
goes from ~26s to ~11s.