@MasterDuke17 2020-09-23
> Maybe we can plan a switch to GMP for the future?

I'm not aware of any technical or licensing impossibility, but it would likely be a lot of grunt work.