@MasterDuke17 2020-09-23
15:02 | jnthn | MasterDuke:  We ended up with libtommath partly out of license concerns (although  LGPL is probably OK; I always forget exactly what it enforces), but in a  bigger part because of libtommath being far easier to build across  platforms.
15:03 | jnthn | MasterDuke: Of note, stock gmp doesn't build on the MSVC compiler.
15:03 | jnthn | There is these days a fork that does: http://mpir.org/