@AlexDaniel 2020-09-24
In other words there wasn't any strong justification, more like an arranged marriage and we just ended up with libtommath.

Just to get a feel of both, I wrote these two programs:

```c
#include <gmp.h>
#include <stdio.h>

mpz_t number1;
mpz_t number2;


int main() {
    mpz_init_set_ui(number1, 0);
    mpz_init_set_ui(number2, 1);

    for (size_t i = 0; i < 10000000; i++) {
        mpz_add(number1, number1, number2);
    }

    printf("heh %lu\n", mpz_get_ui(number1));
    printf("heh %lu\n", mpz_get_ui(number2));

    return 0;
}
```

```c
#include <tommath.h>
#include <stdio.h>

mp_int number1;
mp_int number2;

int main() {
    if (mp_init_ul(&number1, 0) != MP_OKAY)
        printf("whoopsie!");

    if (mp_init_ul(&number2, 1) != MP_OKAY)
        printf("whoopsie!");

    for (size_t i = 0; i < 10000000; i++) {
        if (mp_add(&number1, &number2, &number1) != MP_OKAY)
            printf("whoopsie!");
    }

    printf("heh %u\n", mp_get_u32(&number1));
    printf("heh %u\n", mp_get_u32(&number2));

    return 0;
}
```

Basically, just init two bigints to 0 and 1, and then keep adding the second one to the first one 10_000_000 times. Nothing exciting, but here are the timings:

GMP
```
heh 100000000
heh 1

real	0m0.984s
user	0m0.981s
sys	0m0.001s
```

LibTomMath
```
heh 100000000
heh 1

real	0m1.331s
user	0m1.330s
sys	0m0.001s
```

The error checking with libtommath is required with `-Werror`, but I also tried without it and saw no difference in timings.

I do not know what to think of this, microbenchmarks are kinda nasty. But if basic bigint arithmetic is 1.35x as fast with GMP, then the grunt work may as well be justified.

Also, I think GMP supports rational numbers? Mmmm.