@lizmat 2020-09-24
> In other words there wasn't any strong justification

Again, this remark is not needed and unnecessarily abrasive (again).  If you read the conversation a bit further on IRC, you'd see that @jnthn was working on Windows at the time, using the MSVC compiler.  Not being supported on the compiler that you're using, would be an **excellent** reason for not selecting that library.