@jnthn 2020-09-24
> In other words there wasn't any strong justification, more like an arranged marriage and we just ended up with libtommath.

@AlexDaniel Not building on a major platform / toolchain - especially when it was, at the time, the primary development toolchain of the MoarVM founders - seems strong enough as a reason to pick an alternative that did satisfy that constraint. It's not like it committed MoarVM to always use it, if the situation changed (as it seems to now have done so).

As @lizmat points out, your remark was also unnecessarily abrasive and unconstructive. I've had quite enough of behaviors like:

* Painting anybody who made a decision differently than you imagine you would have under the same constraints as doing things without thought
* Using terms like "naysayer" to describe people who present valid concerns about something you'd like to see happen
* Using loaded or offensive language, both on IRC and in issues
* Disrespectfully dismissing points put forward by others with responses like "this is not about having great debating skills"

Which have sucked much of the joy out of contributing to Raku for a number of people by this point. For now, you continue to get away with this behavior in a Raku context, and ultimately it shall be for the Raku Steering Council to decide if it wishes to continue to entertain it.

So far as MoarVM goes, it's *my* call to make. Please refrain from such behavior; a further instance in the context of MoarVM development shall be grounds for banning you from this project.