@AlexDaniel 2020-09-24
@jnthn I am for honest and direct language that doesn't mislead readers and in fact calls for action to change the situation. The idea of picking a specific library for the project just because you used MSVC at the time isn't a good justification, and there is no need to double down. Raku project is riddled with decisions that were seemingly random at the time but are now extremely hard to undo/redo, here we are just scratching the surface. How about starting with “yeah, maybe it would've been better if we did X, we can perhaps still do it” instead of providing a long description with multiple reasons that have no relevance today. Especially if you want to insist that everything should be *your* call, except that you're making an extremely small amount of said calls.

Granted, in [full log](https://colabti.org/irclogger/irclogger_log/moarvm?date=2020-09-23#l63) you did say that maybe it should be done, but you still seem to be implying that MSVC is a hard requirement (as if… better performance is less important?):


> jnthn | **Given that exists**, we maybe can think about a move to gmp, though of course it needs a vict^Wvolunteer...

Meanwhile, MSVC support doesn't even [seem to be](https://github.com/JuliaLang/julia/issues/12056#issuecomment-487768933) planned in Julia.


> Which have sucked much of the joy out of contributing to Raku for a number of people by this point.

Raku project had 20 years of happy mindless progress in no particular direction. Unsurprisingly, it didn't get anywhere we all want it to be. Accepting the current situation is not a pleasant act. We can't always be happy, it is life. IMO the biggest question for the council right now is what exactly Raku should be and *how can we get there*, but you can of course spend the energy on trying to exclude me because you don't like the things I say and how I say them directly and openly. By the way, I clearly expressed my frustration with the current state of things in [my nomination](https://github.com/Raku/Raku-Steering-Council/blob/main/nominations/2020/AlexDaniel.md). 


![image](https://user-images.githubusercontent.com/5507503/94172710-7f754800-fe9b-11ea-954e-a7ae52e772f6.png)